feat: clean up and extraction of component for view builder
